feat: add SaveLocationForm component with validation and handling for adding/editing locations
